+++ sui_core/src/lib.rs
@@ -7,6 +7,7 @@ pub mod authority_aggregator;
 pub mod authority_batch;
 pub mod authority_client;
 pub mod authority_server;
+pub mod checkpoints;
 pub mod consensus_adapter;
 pub mod execution_engine;
 pub mod full_node;
@@ -16,4 +17,3 @@ pub mod make;
 pub mod safe_client;
 pub mod sui_json;
 pub mod transaction_input_checker;
-pub mod checkpoints;
